bootcamp_week7/bootcamp_day5/requirements.txt
fastapi
uvicorn
redis
faiss-cpu
numpy
pydantic
requests
sentence-transformers
python-dotenv
psycopg2-binary
streamlit
